Conversion from WeeklyTimeRange to AbsoluteTimeRange
The start and end data as stored in the members is transferred into tm structs, which are then converted into timestamps using mktime(). An AbsoluteTimeRange object is instantiated with those timestamps and returned.

Most data in the tm structs is the same as the timestamp of the Sunday which it is relative to; the only thing that changes are the month and day - these are incremented accordingly, depending on the week day.