I'm Trying^tm
The main function of the Appointment Release is constructed. It walks through the creation of an Appointment object by a Patient object, the rescheduling of an App object by a Patient, and the rescheduling of an App by a Clinician.

These are not designed for user input or selection but merely to prove that these objects can be instantiated by the Classes That Be.

NOTE: This is not actually compiling yet lol.

Other changes:
-Constructor for CPSICUser
-Constructor for Patient
-Constructor for Clinical
-Creation of a new class "MedicalRecord" to be an object held by a Patient. It contains the health records and the various boolean stats that were previously directly members.
-I think that's it.